refactor: moved comparison operators enum from global to atomized file per discussion with GriffinR
