+++ include/constants/comparison_operators.h
@@ -1,0 +1,15 @@
+#ifndef GUARD_CONSTANTS_COMPARISON_OPERATORS_H
+#define GUARD_CONSTANTS_COMPARISON_OPERATORS_H
+
+enum ComparisonOperators
+{
+    LESS_THAN,
+    EQUAL,
+    GREATER_THAN,
+    LESS_THAN_OR_EQUAL,
+    GREATER_THAN_OR_EQUAL,
+    NOT_EQUAL,
+    COMPARISON_OPERATORS_COUNT,
+};
+
+#endif // GUARD_CONSTANTS_COMPARISON_OPERATORS_H
